requirements.txt
flask
langchain
langchain-community
langchain-google-genai
python-dotenv
PyMuPDF
